refactor(format): extract export redaction owner
